requirements.txt
# Aura AI Backend — Python Dependencies
# Install: pip install -r requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# AI client (REST via httpx)
httpx==0.27.0

# Astrology calculations (Swiss Ephemeris)
# Optional but recommended for accurate planetary positions
pyswisseph==2.10.3.2

# Data validation
pydantic==2.9.0

# Environment variables
python-dotenv==1.0.1

# Database (optional: Supabase)
supabase==2.7.0
asyncpg==0.29.0

# Payments
# razorpay==1.4.2    # Uncomment when using SDK (REST used by default)

# CORS & middleware
python-multipart==0.0.9

# Image processing (for Receipts Judge)
Pillow==10.4.0
gunicorn==21.2.0
